feat(curated): DB-authoritative curated lists — registry inversion, membership snapshot, seed cutover
- @tokens/asset-registry/curated-lists: single source for curated slugs,
  order, aliases, reserved community-list slugs, and fallback names
  (replaces four drifted copies; admin can no longer strip lsts membership)
- registry data files build canonical assets from data/list-mints.ts;
  curated-token-lists.ts is a deprecated shim (CANONICAL_ASSETS verified
  byte-identical); collections.ts deleted (no importers)
- cloudrun-assets: effective curated-membership snapshot source (asset-level
  membership expanded to active sibling variants, yield/LST mints routed only
  to lsts, live 'all' union, ~60s TTL, stale-forever, cold load throws) served
  via new curatedMembershipGetSnapshot RPC and wired into cron/cacheWarm/
  trending consumers
- nightly seed no longer writes collection membership (identity only);
  guarded seed-curated-collections-fixture bootstraps fresh DBs from a frozen
  asset-level fixture; one-off cleanup job for the materialized 'all' rows
- adminSeedAsset: membership is additive/multi-list (exclusivity throw gone)
- apps/web reads slugs from the leaf module; client-side membership arrays
  (TOKEN_SEARCH_CATEGORIES, STATIC_LST_MINTS) deleted
- curation text-rewrite scripts hard-deprecated; validate/audit/upsert-ondo
  retargeted to list-mints.ts
