+++ apps/api/src/app/api/v2/lists/check-slug/route.ts
@@ -4,16 +4,14 @@ import { route } from '@/effect/next-route';
 import { BadRequestError } from '@tokens/effect';
 import { tokenListsGetBySlug } from '@/lib/cloudrun';
 
-import { normalizeCuratedSlug } from '../_shared';
+import { isReservedListSlug } from '@tokens/asset-registry/curated-lists';
 
 /**
- * Mirrors the write-side rules in cloudrun-assets `tokenListsMutations`
- * (TOKEN_LIST_SLUG_REGEX / isReservedTokenListSlug). Duplicated rather than
- * shared because the API and the assets service are separate deployables —
- * keep the two in sync when either changes.
+ * Mirrors the write-side slug regex in cloudrun-assets `tokenListsMutations`
+ * (TOKEN_LIST_SLUG_REGEX). Reservation rules are shared via
+ * `@tokens/asset-registry/curated-lists` so the two deployables cannot drift.
  */
 const SLUG_REGEX = /^[a-z][a-z0-9-]{2,62}$/;
-const RESERVED_SEGMENTS = new Set(['all', 'lists', 'curated', 'tokens', 'search-tokens', 'check-slug']);
 
 export type SlugUnavailableReason = 'invalid' | 'reserved' | 'taken';
 
@@ -40,7 +38,7 @@ export const GET = route(
             const unavailable = (reason: SlugUnavailableReason) => ({ slug: raw, available: false, reason });
 
             if (!SLUG_REGEX.test(raw)) return unavailable('invalid');
-            if (RESERVED_SEGMENTS.has(raw) || normalizeCuratedSlug(raw) !== null) return unavailable('reserved');
+            if (isReservedListSlug(raw)) return unavailable('reserved');
 
             const existing = yield* tokenListsGetBySlug({ slug: raw });
             if (existing) return unavailable('taken');
