+++ packages/asset-registry/src/index.ts
@@ -56,5 +56,17 @@ export { getHub, getHubByMint, listHubs } from './hubs';
 export type { VariantHub } from './variant-hubs';
 export { getVariantHubByAssetId, getVariantHubById, getVariantHubByMint, listVariantHubs } from './variant-hubs';
 
-export type { AssetCollection } from './collections';
-export { getCollection, listCollectionMembers, listCollections } from './collections';
+export type { CuratedListSlug } from './curated-lists';
+export {
+    ADMIN_ASSIGNABLE_CURATED_SLUGS,
+    ALL_PSEUDO_SLUG,
+    CURATED_LIST_FALLBACK_NAMES,
+    CURATED_LIST_ORDER,
+    CURATED_LIST_SLUGS,
+    CURATED_SLUG_ALIASES,
+    HOME_CATEGORY_SLUGS,
+    STATIC_RESERVED_LIST_SLUGS,
+    isCuratedListSlug,
+    isReservedListSlug,
+    normalizeCuratedListSlug,
+} from './curated-lists';
