+++ scripts/audit-asset-registry-hubs.ts
@@ -1,6 +1,24 @@
 /* eslint-disable no-console */
 
-import { CURATED_TOKEN_LISTS } from '../packages/asset-registry/src/data/curated-token-lists.ts';
+import {
+    CURRENCY_MINTS,
+    ETF_MINTS,
+    LST_MINTS,
+    MAJORS_MINTS,
+    METALS_MINTS,
+    RWA_MINTS,
+    STOCKS_MINTS,
+} from '../packages/asset-registry/src/data/list-mints.ts';
+
+const REGISTRY_LIST_MINTS = [
+    { id: 'majors', addresses: MAJORS_MINTS },
+    { id: 'lsts', addresses: LST_MINTS },
+    { id: 'currencies', addresses: CURRENCY_MINTS },
+    { id: 'rwas', addresses: RWA_MINTS },
+    { id: 'etfs', addresses: ETF_MINTS },
+    { id: 'metals', addresses: METALS_MINTS },
+    { id: 'stocks', addresses: STOCKS_MINTS },
+];
 import { MINT_HOME_OVERRIDES } from '../packages/asset-registry/src/data/mint-home-overrides.ts';
 import { getHubByMint } from '../packages/asset-registry/src/hubs.ts';
 import { getVariantHubById } from '../packages/asset-registry/src/variant-hubs.ts';
@@ -14,7 +32,7 @@ function main(): void {
     const errors: string[] = [];
 
     // 1) All curated mints resolve.
-    for (const list of Object.values(CURATED_TOKEN_LISTS)) {
+    for (const list of REGISTRY_LIST_MINTS) {
         if (!list || typeof list !== 'object') continue;
         const listId = String((list as { id?: unknown }).id ?? '');
         const addresses = Array.isArray((list as { addresses?: unknown }).addresses)
